Issue 692: Added enableNodeLevelSecrets attribute to secretBinding in deployment patterns and policies. Renamed a few test source packages.
