+++ src/main/scala/org/openhorizon/exchangeapi/table/deploymentpattern/OneSecretBindingService.scala
@@ -4,4 +4,5 @@ final case class OneSecretBindingService(serviceOrgid: String,
                                          serviceUrl: String,
                                          serviceArch: Option[String],
                                          serviceVersionRange: Option[String],
-                                         secrets: List[Map[String, String]])
+                                         secrets: List[Map[String, String]],
+                                         enableNodeLevelSecrets: Option[Boolean] = Option(false))
